Add support for export object with aliases.
Previously when faced with `exports = {...}` we did
a check on whether the object is of the form `{A, B, C}`
and move the `export` keywords.

However, for aliases like `{Alias: B}` we bailed into
the creating exports object mode, emitting:

```
export const obj = {Alias: B};
```
This is undesirable because it requires changes on the import side.

For objects that only contain aliases there is a better answer, emit the
aliases as - `export {B as Alias}`. This change does that.

PiperOrigin-RevId: 314967742