Always move static members into a class declaration
Before this commit class properties declared outside of the class definition
would be moved up into the class definition iff the member had a type in its
JSDoc. However, we can still move a member into a function even if we don't
know its type, and just attach an `any` later on.

For "@const" (readonly) properties, we also the property initializer if the
original initializer was in a top-level scope. Ideally we would drop the
"@const" attribute if it was present but the original intialization was not in a
top-level scope; this could be done in a separate commit, but I imagine it's not
a very common pattern.

PiperOrigin-RevId: 315706628