Elide explicit "any" on class members when initializer is present
If the initializer for a class member is present, the TypeScript typechecker
will probably be able to infer a narrower type than any. In most cases this is
more useful for users as it avoids having to manually narrow the type later on.

I say "most cases" because there are some cases where closure JS translated in
this way won't compile. For example,

```
class C {}

C.o = {};
C.o.n = 1;
```

will be translated to

```
class C {
  static o = {};
}

C.o.n = 1;
```

which fails to compile because the type of `C.o` is `{}`.

In general I think this change is better, though, because it is easier to add an
"any" to gents-generated TS than remove one.

PiperOrigin-RevId: 315979195