Improve strategy for determining if an imported module is TS
Today gents renames imported modules whose file path starts with a particular
prefix (default "google3") to the name they would have if they were a TS source.
But this is not always accurate, and there is a better way to try to figure out
if a module comes from a TS file: since we know the name of the file providing
the module, we check if it ends with .closure.js and is in a "generated files"
directory. If both of these conditions are true, the module was likely produced
by a TS source.

We also make a special case for imported files generated from the sources we are
converting -- these files are not yet TS, but will be after we finish conversion
of the passed TS source. However, we do not actually convert the generated files
in gents. So, we just treat the generated files as having TS import names, since
they will after conversion is complete.

PiperOrigin-RevId: 321166872