Destruct default export when keys are valid JS identifiers
Prior to this commit, gents would destruct a default object export only when the
all the values in the export referenced local identifiers. This meant that

```js
class A {}
function foo() {}
exports = {A, foo: foo}
```

would be converted to

```ts
export class A {}
export function foo() {}
```

but more complex expressions, like

```js
exports = {pi: 22 / 7}
```

would be converted to a named object export:

```ts
export obj = {pi: 22 / 7}
```

The downside of this named object export is that all importers of properties in
the previously-default export must be updated, unlike in the first case where
the export object is discarded.

This commit adds support for destructing export objects with non-identifier
export expressions to named exports. For example, `exports = {pi: 22 / 7}`
becomes

```ts
export const pi = 22 / 7;
```

---

There are some caveats and important notes here. The first is that this
destructuring is only legal when the keys of the exports object are valid JS
identifier. For example, no key in

```
exports = {
  'abc-def': 0,
  [123]: 123,
  [foo()]: foo(),
}
```

can be raised to a JS identifier, so we have to fallback on the object export
here. This is not a big problem, because any consumer could not import one of
these exports by name anyway.

The second note is that there may be subtle semantic differences between, for
example, `exports = {A: 0}` and `exports.A = 0` due to how JScomp handles
mutable exports. This means that raising `exports = {A: <expr>}` to `export
const A = <expr>` may break some code.

However, we predict the occurence of such cases to be relatively low, and it is
worth trying this raising to see where/how it breaks so that we have real
examples of why we can't use it.

PiperOrigin-RevId: 321409790