Handle optional binding patterns
Today gents throws when trying to mark a binding pattern as optional, for
example in a code like

```
/** @param {{n: (string|undefined) }=} */
function f({n}) {}
```

because `{n}` is not a named parameter, and currently all optional items are
expected to be names.

This commit fixes the error by handling the case where the optional item is a
binding pattern, and emitting `= {}` as a marker that the item is optional. E.g.
the code above is emitted as

```
function f({n} = {}) {}
```

Note that we cannot emit `?`, as this operator only applies to names.

PiperOrigin-RevId: 321814427