Only collect objects in an exported block when it belongs to `exports`
Gents raises exports declared in an `exports = {}` block to constants on the
source file that are exported with the TS `export` keyword. However, currently
we try to collect such exports even for named exports that have an object value.
For example, we would try to do this for an export like

```
/** @enum {number} */
exports.N = {
  One: 1,
}
```

but this is incorrect, because the export `N` as a whole should be recorded, not
its enum properties.

PiperOrigin-RevId: 322245547