Convert construct signatures and this annotations in function types
This commit adds support for properly converting construct signatures and `this`
annotations in function type annotations.

## Construct signatures

For a code like

```js
/** @param {function(new: Foo, ...)} ctor */
function makeFoo(ctor) { return new ctor(); }
```

`ctor` is a `new`-able function that returns a `Foo`. Currently this is
converted to

```ts
function makeFoo(ctor: (...p1: any[]) => any) { return new ctor(); }
```

but should be converted to a TypeScript construct signature, as

```ts
function makeFoo(ctor: new (...p1: any[]) => Foo) { return new ctor(); }
```

## `this` annotations

In closure, the type of `this` bound in a function can be specified in the
function type, for example like

```js
/**
 * @param {function(this: Foo, number)} callThroughFoo
 * @param {number} num
 */
function deferToFoo(callThroughFoo, num) { callThroughFoo(num); }
```

Previously the `this` implicit parameter was ignored, perhaps because in earlier
versions of TS it wasn't supported, but [since TS 2.0](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-0.html#specifying-the-type-of-this-for-functions)
it has been possible to do so. That is, the above function should be converted to

```ts
function deferToFoo(
  callThroughFoo: (this: Foo, p1: number) => void, num: number) {
  callThroughFoo(num);
}
```

Which this commit enables.

PiperOrigin-RevId: 322245883