Fix ordering of function keywords - private/protected/public, abstract, static and async.
Certain combinations of protected/static/async/abstract are forbiden in TS.
This test adds the maximal combinations:
- abstract protected async
- static protected async

Also, fixes the emit for the 'abstract' case. Because we need to drop the body
of abstract, we have custom emit there, which was missing handling of some
keywords. All keywords should be supported now.

PiperOrigin-RevId: 322843127