Correctly emit "void" in union return types
Currently gents emits the "void" type only in the context of a return type. When
the type to be returned is a union type, we construct the union by recursively
building the types of the union, but prior to this commit we would not pass down
information that the union subtype is located in a return context.

This commit fixes that by passing down information about whether the union is in
a return context when generating the subtypes.

PiperOrigin-RevId: 324294067