introduce overload of GoogPromise.resolve()
When we fixed the return type of Promise.resolve in general, we left
behind GoogPromise.resolve as returning `any`.  Fixing this right requires
an extra batch of changes, because of an API quirk: some callers want a `GoogPromise<void>`.  The underlying API uses TTL to express a similar
conditional type.

So we will:
1. Introduce an overload that returns `void`, submit that.
2. Change calling code to either use the new void-returning API, or
   the proper new type.
3. Change the return type now that callers are fixed.

We need to do Clutz changes on both sides of #2 because we want to
expose the new overload before touching callers, and we need to fix
callers before fixing the return type.

PiperOrigin-RevId: 325037277