Default to "any" for record fields missing type declaration
Currently if the type of record field is not declared when converting type
annotations, we leave the field type null, causing the next branch to blow up
because it expects the field type to be defined. Since the field type in such
cases is implicit any anyway, use "any" as the default type rather than leaving
it as null, preventing a null dereference and providing better types.

PiperOrigin-RevId: 325051310