unify the clutz type of `Promise.resolve()`
Clutz special-cases the types of various Promise-related functions.
It had a further special exception to produce a wrong type for
specifically the goog.Promise type found in Closure library, as a
backwards compat thing.

In this CL, I delete the exception to make all Promise.resolve()s
use proper types.

PiperOrigin-RevId: 328568498